Changed the way of resetting db in tests for cached dbs

+++ src/tmp_dir.cpp
@@ -55,6 +55,13 @@ TmpDir::TmpDir(std::string prefix)
     boost::filesystem::create_directories(this->path);
 }
 
+TmpDir& TmpDir::operator=(TmpDir&& other)
+{
+    this->path = other.path;
+    other.path = "";
+    return *this;
+}
+
 void TmpDir::Execute(std::string exe, std::string args) const
 {
     if(miopen::IsEnabled(MIOPEN_DEBUG_SAVE_TEMP_DIR{}))
@@ -70,7 +77,8 @@ TmpDir::~TmpDir()
 {
     if(!miopen::IsEnabled(MIOPEN_DEBUG_SAVE_TEMP_DIR{}))
     {
-        boost::filesystem::remove_all(this->path);
+        if(!this->path.empty())
+            boost::filesystem::remove_all(this->path);
     }
 }
